+++ packages/components/src/organisms/ProductShelf/index.ts
@@ -1,0 +1,8 @@
+export { default } from './ProductShelf'
+export type { ProductShelfProps } from './ProductShelf'
+
+export { default as ProductShelfItems } from './ProductShelfItems'
+export type { ProductShelfItemsProps } from './ProductShelfItems'
+
+export { default as ProductShelfItem } from './ProductShelfItem'
+export type { ProductShelfItemProps } from './ProductShelfItem'