+++ packages/core/src/components/region/RegionButton/RegionButton.tsx
@@ -9,7 +9,6 @@ function RegionButton() {
 
   return (
     <UIButton
-      data-fs-regionalization-button
       variant="tertiary"
       size="small"
       icon={<Icon name="MapPin" width={24} height={24} />}