+++ packages/core/src/components/sections/Hero/Hero.tsx
@@ -1,50 +1,49 @@
-import type {
-  HeroHeaderProps as UIHeroHeaderProps,
-  HeroImageProps as UIHeroImageProps,
-  HeroProps as UIHeroProps,
-} from '@faststore/ui'
+import type { HeroProps as UIHeroProps } from '@faststore/ui'
 import {
   Hero as UIHero,
   HeroHeader as UIHeroHeader,
   HeroImage as UIHeroImage,
 } from '@faststore/ui'
-
+import { ReactNode } from 'react'
 import { Image } from 'src/components/ui/Image'
-
 import Section from '../Section'
+
 import styles from './section.module.scss'
 
-export type HeroProps = UIHeroProps &
-  UIHeroHeaderProps &
-  UIHeroImageProps & {
-    /**
-     * Specifies the image URL.
-     */
-    imageSrc: string
-    /**
-     * Alternative description of the image.
-     */
-    imageAlt: string
+export type HeroProps = {
+  title: string
+  subtitle: string
+  link?: {
+    text: string
+    url: string
   }
+  image: {
+    src: string
+    alt: string
+  }
+  variant?: UIHeroProps['variant']
+  colorVariant?: UIHeroProps['colorVariant']
+  icon?: ReactNode
+}
 
 const Hero = ({
-  icon,
   link,
   title,
   subtitle,
-  linkText,
-  imageAlt,
-  imageSrc,
+  image,
+  variant = 'primary',
+  colorVariant = 'main',
+  icon,
 }: HeroProps) => {
   return (
     <Section className={`${styles.section} section-hero`}>
-      <UIHero>
+      <UIHero colorVariant={colorVariant} variant={variant}>
         <UIHeroImage data-fs-hero-image>
           <Image
             loading="eager"
             priority
-            src={imageSrc}
-            alt={imageAlt}
+            src={image.src}
+            alt={image.alt}
             width={360}
             height={240}
             sizes="(max-width: 768px) 70vw, 50vw"
@@ -53,8 +52,8 @@ const Hero = ({
         <UIHeroHeader
           title={title}
           subtitle={subtitle}
-          link={link}
-          linkText={linkText}
+          link={link?.url}
+          linkText={link?.text}
           icon={icon}
         />
       </UIHero>