+++ CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.111-alpha.0](https://github.com/vtex/faststore/compare/v2.0.110-alpha.0...v2.0.111-alpha.0) (2023-04-18)
+
+### Features
+
+- Adds `ProductGrid` component ([#1719](https://github.com/vtex/faststore/issues/1719)) ([9cbe71f](https://github.com/vtex/faststore/commit/9cbe71f81395297c14b258efdd336b00f6b71071))
+
+## [2.0.110-alpha.0](https://github.com/vtex/faststore/compare/v2.0.109-alpha.0...v2.0.110-alpha.0) (2023-04-18)
+
+### Features
+
+- Add `ProductShelf` component ([#1713](https://github.com/vtex/faststore/issues/1713)) ([6a387a6](https://github.com/vtex/faststore/commit/6a387a665aa8a1eb65122abcce67d82b9ead4234))
+
+## [2.0.109-alpha.0](https://github.com/vtex/faststore/compare/v2.0.108-alpha.0...v2.0.109-alpha.0) (2023-04-18)
+
+### Features
+
+- Apply adjusts in `HeroSection` and `BannerText` section ([#1715](https://github.com/vtex/faststore/issues/1715)) ([968f95a](https://github.com/vtex/faststore/commit/968f95afb808c104a795e078f45c37ca619d4697))
+
+## [2.0.108-alpha.0](https://github.com/vtex/faststore/compare/v2.0.107-alpha.0...v2.0.108-alpha.0) (2023-04-17)
+
+### Features
+
+- Adds `EmptyState` component ([#1716](https://github.com/vtex/faststore/issues/1716)) ([82eaf9a](https://github.com/vtex/faststore/commit/82eaf9a628f417ee58d7cacb2678f8cdd81bfaec))
+
+## [2.0.107-alpha.0](https://github.com/vtex/faststore/compare/v2.0.106-alpha.0...v2.0.107-alpha.0) (2023-04-17)
+
+### Chores
+
+- Removes `Form` component ([#1717](https://github.com/vtex/faststore/issues/1717)) ([ffe1c2e](https://github.com/vtex/faststore/commit/ffe1c2ec93f261e24bf3c5a295cf359cf53bcc63))
+
+## [2.0.106-alpha.0](https://github.com/vtex/faststore/compare/v2.0.105-alpha.0...v2.0.106-alpha.0) (2023-04-12)
+
+### Features
+
+- Adds `RegionModal` component ([#1710](https://github.com/vtex/faststore/issues/1710)) ([d3b6c36](https://github.com/vtex/faststore/commit/d3b6c36af3fe3d9c826884654c8708d4a1d14d2d))
+
+## [2.0.105-alpha.0](https://github.com/vtex/faststore/compare/v2.0.104-alpha.0...v2.0.105-alpha.0) (2023-04-12)
+
+### Features
+
+- add an `user-agent` header to `fetchAPI` method ([#1708](https://github.com/vtex/faststore/issues/1708)) ([e1714c3](https://github.com/vtex/faststore/commit/e1714c350ef2a5c1a30b1a46c4e79d4a8d24993b)), closes [#1709](https://github.com/vtex/faststore/issues/1709)
+
 ## 2.0.104-alpha.0 (2023-04-10)
 
 **Note:** Version bump only for package faststore